register setting and add IT
